Upgrade to Rails 6.1
This upgrade is needed for the upcoming work.

In the process of upgrading the following error was
surfaced while running the unit tests:

Failure/Error: <%= stylesheet_link_tag "application", :media => "all" %>
  ActionView::Template::Error:
    Error: "calc(0px)" is not a number for `max'
            on line 1261:20 of stdin, in function `max`
            from line 1261:20 of stdin
    >> @supports (margin: max(calc(0px))) {
        -------------------^

Read more about it:
alphagov/govuk-frontend#1350

This issue goes away if the assets are precompiled. We do not, however,
want to push the precompiled assets on our branches, so turning on the
assets compilation phase for dev and test envs solves the issue.

Also setting config.assets.css_compressor to nil (check the link above).