Dockerise admin app
Adds the circle ci plumbing, Dockerfile and application changes than
enable a dockerised build of the admin app.

See: https://transformuk.atlassian.net/browse/HOTT-1216

This is basically so that we can immutably revert/deploy and manage our
infrastructure and makes this project consistent with the other ones we
manage.