HOTT-1223 Fixed detection of empty release notes
Turns out `wc -l` counts explicit line endings, because the last line of git log output lacks a new line (or bash is stripping it when writing to a file) this means wc -l is returning the actual line count - 1.

Which means if there was only a single PR merged, then this will be read as 0 lines and the fallback release notes will be created.

Solution is to just rely on bash' ability to detect empty files.